🧹 Extract useToastNotifications to separate file
